feat: rewrite service and controller with try catch

+++ src/modules/doctor/doctor.service.ts
@@ -1,6 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import CreateDoctorDto from './dto/create-doctor.dto';
 import Doctor from './entity/doctor.entity';
 
 @Injectable()
@@ -10,19 +11,43 @@ export default class DoctorService {
     private doctorRepository: Repository<Doctor>,
   ) {}
 
-  createDoctor(doctor: Doctor): Promise<Doctor> {
-    return this.doctorRepository.save(doctor);
+  async createDoctor(doctorDto: CreateDoctorDto): Promise<CreateDoctorDto> {
+    try {
+      const doctor = await this.doctorRepository.create(doctorDto);
+      await this.doctorRepository.save(doctor);
+      return doctor;
+    } catch (err) {
+      throw new HttpException(`${err}`, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
   }
 
-  findAll(): Promise<Doctor[]> {
-    return this.doctorRepository.find();
+  async getAllDoctors(): Promise<Doctor[]> {
+    try {
+      const doctors = await this.doctorRepository.find();
+      return doctors;
+    } catch (err) {
+      throw new HttpException(`${err}`, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
   }
 
-  findOne(id: number): Promise<Doctor | null> {
-    return this.doctorRepository.findOneBy({ id });
+  async getDoctorByID(id: number): Promise<Doctor | null> {
+    try {
+      const doctor = await this.doctorRepository.findOne({ where: { id } });
+      if (!doctor) throw new Error(`Can't find doctor with id ${id}`);
+      return doctor;
+    } catch (err) {
+      throw new HttpException(`${err}`, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
   }
 
-  async remove(id: number): Promise<void> {
-    await this.doctorRepository.delete(id);
+  async deleteDoctorById(id: number): Promise<string> {
+    try {
+      const doctor = await this.doctorRepository.findOne({ where: { id } });
+      if (!doctor) throw new Error(`Can't find doctor with id ${id}`);
+      await this.doctorRepository.delete(id);
+      return `Doctor with id ${id} was deleted`;
+    } catch (err) {
+      throw new HttpException(`${err}`, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
   }
 }
